+++ python/packages/durabletask/agent_framework_durabletask/_workflows/activity.py
@@ -25,13 +25,13 @@ from agent_framework import Executor, Workflow, WorkflowEvent
 from agent_framework._workflows._runner_context import YieldOutputEventType
 from agent_framework._workflows._state import State
 
-from ._workflow_orchestrator import (
+from .orchestrator import (
     SOURCE_HITL_RESPONSE,
     SOURCE_ORCHESTRATOR,
     execute_hitl_response_handler,
 )
-from ._workflow_runner_context import CapturingRunnerContext
-from ._workflow_serialization import deserialize_value, serialize_value
+from .runner_context import CapturingRunnerContext
+from .serialization import deserialize_value, serialize_value
 
 
 def execute_workflow_activity(executor: Executor, input_json: str, workflow: Workflow | None = None) -> str:
