+++ python/packages/durabletask/agent_framework_durabletask/_workflows/orchestrator.py
@@ -48,8 +48,8 @@ from agent_framework._workflows._edge import (
 )
 from agent_framework._workflows._state import State
 
-from ._workflow_context import WorkflowOrchestrationContext
-from ._workflow_serialization import (
+from .context import WorkflowOrchestrationContext
+from .serialization import (
     deserialize_value,
     reconstruct_to_type,
     resolve_type,
